Partial Competition Implementation

+++ Assets/Scripts/EvolutionScripts/TournamentCandidate.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class TournamentCandidate
 {
     public AgentVillagerEvolution candidate;
-    public int fitness;
-    public TournamentCandidate(AgentVillagerEvolution agent, int fitness)
+    public float fitness;
+    public TournamentCandidate(AgentVillagerEvolution agent, float fitness)
     {
         candidate = agent;
         this.fitness = fitness;
